# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "System manage music",
        "title": "Music API REST",
        "termsOfService": "https://juancx.site/",
        "contact": {
            "name": "API Support",
            "email": "juanm.campos@unas.edu.pe"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:4021",
    "basePath": "/",
    "paths": {
...
        "/api/v1/artists": {
            "put": {
                "description": "Método que permite Actualiza una instancia del objeto Artists en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Actualiza una instancia de Artists",
                "parameters": [
                    {
                        "description": "Datos para actualizar Artists",
                        "name": "RequestArtists",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/artists.RequestArtists"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    }
                }
            },
            "post": {
                "description": "Método que permite crear una instancia del objeto Artists en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Crear una instancia de Artists",
                "parameters": [
                    {
                        "description": "Datos para crear Artists",
                        "name": "RequestArtists",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/artists.RequestArtists"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    }
                }
            }
        },
        "/api/v1/artists/": {
            "get": {
                "description": "Método que permite obtener todas las instancias del objeto Artists en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Obtiene todas las instancias de Artists",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    }
                }
            }
        },
        "/api/v1/artists/{id}": {
            "get": {
                "description": "Método que permite obtener una instancia del objeto Artists en la base de datos por su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Obtiene una instancia de Artists por su id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artists ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    }
                }
            },
            "delete": {
                "description": "Método que permite eliminar una instancia del objeto Artists en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Elimina una instancia de Artists",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artists ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/artists.ResponseArtists"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "artists.RequestArtists": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "artists.ResponseArtists": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}